Added classes exceptions, tests for it, method searchById

+++ src/main/java/Shop/manager/ProductManager.java
@@ -2,6 +2,8 @@ package Shop.manager;
 
 import Shop.shopsProducts.Product;
 import Shop.repository.ProductRepository;
+import Shop.negativeExceptions.NotFoundException;
+import Shop.negativeExceptions.DuplicateException;
 
 public class ProductManager {
     private ProductRepository repository;
@@ -15,14 +17,26 @@ public class ProductManager {
     }
 
     public void add(Product newProduct) {
-        repository.save(newProduct);
+        if (searchById(newProduct.getId()) == null) {
+            repository.save(newProduct);
+        }else{
+            throw new DuplicateException(
+                    "Element with id: " + newProduct.getId() + " has already exists!"
+            );
+        }
     }
 
     public void deleteById(int id) {
-        repository.deleteById(id);
+        if (searchById(id) == null) {
+            throw new NotFoundException(
+                    "Element with id: " + id + " not found!"
+            );
+        }else{
+            repository.deleteById(id);
+        }
     }
 
-    public Product[] searchBy(String text) {
+    public Product[] searchByName(String text) {
         Product[] result = new Product[0];
         for (Product product: repository.getProducts()) {
             if (matches(product, text)) {
@@ -35,6 +49,15 @@ public class ProductManager {
             }
         }
         return result;
+    }
+
+    public Product searchById(int id) {
+        for (Product product: repository.getProducts()) {
+            if (id == product.getId()) {
+                return product;
+            }
+        }
+        return null;
     }
 
     public boolean matches(Product product, String search) {
